Former-commit-id: 081d47b28c4358c7e4fd045d3577bcd45210ab4c

+++ moa-server/grails-app/controllers/moa/persistence/WebhookController.groovy
@@ -32,7 +32,7 @@ class WebhookController extends RestfulController {
     }
 
     def gnpsUpdate() {
-        def id = params.id as long
+        def id = params.id
         def event = params.event
 
         log.info("Received GNPS update (event = $event) for $id")
